LayoutTests/media/video-loopcount-expected.txt

TEST(video.loopCount == 2) OK
TEST(video.currentLoop == 0) OK
EVENT(ended) TEST(video.currentLoop == 1) OK
END OF TEST
